tcp_bridge/requirements.txt
pexpect==4.8.0
pyserial==3.4
